refactor: toggle molecules => atoms
